feat: 예외 케이스 추가 및 /echo 복호화 api 추가

+++ src/main/java/com/example/postmantestapi/service/MemberServiceImpl.java
@@ -2,10 +2,9 @@ package com.example.postmantestapi.service;
 
 import com.example.postmantestapi.entity.Member;
 import com.example.postmantestapi.repository.MemberRepository;
+import com.example.postmantestapi.response.Response;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -14,28 +13,49 @@ public class MemberServiceImpl implements MemberService {
     private final MemberRepository memberRepository;
 
     @Override
-    public Member login(Member member) {
-        Long id = member.getId();
-        Member byId = memberRepository.findOneById(id);
-        return byId;
+    public Response<String> login(Member member) {
+        Member foundMember = memberRepository.findOneByNickName(member.getNickName());
+        if (foundMember == null) {
+            return new Response<>("0001", "회원 정보가 없습니다.", null);
+        }
+        if (!member.getPassword().equals(foundMember.getPassword())) {
+            return new Response<>("0002", "비밀번호가 맞지 않습니다.", null);
+        }
+        return new Response<>("0000", "로그인 성공", foundMember.getName());
     }
 
     @Override
-    public Member join(Member member) {
-        Member save = memberRepository.save(member);
-        return save;
+    public Response<String> join(Member member) {
+        if (memberRepository.existsByNickName(member.getNickName())) {
+            return new Response<>("0003", "닉네임이 이미 존재합니다.", null);
+        }
+        memberRepository.save(member);
+        return new Response<>("0000", "회원가입 성공", null);
     }
 
     @Override
-    public Member update(Member member) {
-        Member byId = memberRepository.findOneById(member.getId());
-        return memberRepository.save(byId);
+    public Response<String> update(Member member) {
+        Member foundMember = memberRepository.findOneById(member.getId());
+        if (foundMember == null) {
+            return new Response<>("0001", "회원 정보가 없습니다.", null);
+        }
+        if (memberRepository.existsByNickName(member.getNickName())) {
+            return new Response<>("0003", "닉네임이 이미 존재합니다.", null);
+        }
+        foundMember.setName(member.getName());
+        foundMember.setPassword(member.getPassword());
+        foundMember.setNickName(member.getNickName());
+        memberRepository.save(foundMember);
+        return new Response<>("0000", "비밀번호 변경 성공", null);
     }
 
     @Override
-    public Member leave(Member member) {
+    public Response<String> leave(Member member) {
+        Member foundMember = memberRepository.findOneById(member.getId());
+        if (foundMember == null) {
+            return new Response<>("0001", "회원 정보가 없습니다.", null);
+        }
         memberRepository.deleteById(member.getId());
-        return member;
+        return new Response<>("0000", "회원 탈퇴 성공", null);
     }
-
 }
